feat(hydra-js): renderEndpoint config option + server-rendered docs
Promotes what was an Astro-example-internal pattern in commit 1
(f37e4f7) into a first-class bridge feature usable by ANY server-only
frontend - PHP, Django, Rails, Laravel, Symfony, Go templates, plus
Astro. Driven by the user's observation that the diff + POST + swap
mechanics aren't Astro-specific; they're "what every framework without
client-side reactivity needs."

# What changed in the bridge

hydra.src.js now exports `findChangedUnit(prevForm, newForm)` and the
Bridge constructor accepts two new options:

  initBridge({
    page: { ... },
    blocks: { ... },
    renderEndpoint: '/api/render',         // <-- new
    renderContainer: '#content',            // <-- new, optional, default '#content'
  });

When `renderEndpoint` is set the bridge installs an internal callback
that:

  1. Diffs prev vs new formData via findChangedUnit
  2. POSTs { unit, formData } to renderEndpoint
  3. Swaps the returned HTML in:
       - unit.unit === 'page'  -> renderContainer.innerHTML = response
       - unit.unit === 'block' -> [data-block-uid=X].outerHTML = response

Race protection: if a render is already in flight when the next
FORM_DATA arrives, the new render is dropped (the newer formData is
what we diff next time - the dropped one would have been stale anyway).

Error handling: a non-OK response logs but doesn't throw; lastForm is
reset so the next FORM_DATA re-attempts the same change.

If both `renderEndpoint` and `onEditChange` are set, renderEndpoint
wins (documented as a configuration error - we don't try to merge
the semantics).

# Why this lives in hydra.src.js and not a separate package

The diff is pure JS, knows nothing about Astro, and is small (~80 lines
including helpers). Bundling it into the existing hydra.js means PHP /
Django / Rails users get the diff + POST + swap by loading ONE script
and passing ONE config option - they don't need a second package or to
wire glue code per language. Same script that React / Vue users already
load just gains a new capability.

# Astro example simplified

docs/examples/test-astro/src/main.js no longer inlines the diff or POST
logic - it just sets `renderEndpoint: '/api/render'` and lets the bridge
do the work. The local diff.js is deleted; findChangedUnit is the
authoritative version.

# Documentation

New: docs/server-rendered-frontends.md - dedicated guide explaining
when you need the pattern (no client reactivity), how it works
(diff -> POST -> outerHTML swap), the data-block-uid contract, and
worked examples for Astro AND PHP. Plus a one-line render-call table
for Astro / PHP / Django / Rails / Laravel / Symfony / Go showing how
trivial the per-framework adaptation is.

Updated:
  - docs/index.md - toctree entry; framework list mentions PHP/Django/
    Rails/Laravel via the new pattern.
  - docs/architecture.md - the "Switching frontends mid-edit
    (Nuxt -> Next -> Astro)" sentence extended to mention server-only
    frameworks participate via the same protocol + the render-endpoint
    pattern.
  - docs/build-a-frontend.md - admonition at the top routing
    server-only users to the new guide.

# Verification

  cd packages/hydra-js && npm run build  # esbuild bundles cleanly,
                                          # hydra.js is 454.6kb (up
                                          # from prior, marginal)

Test plan unchanged from PR #229 description; this commit doesn't
change the Astro fixture surface, only how it's wired internally.
